# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.Response:
    properties:
      CreationDate:
        type: string
      FormationDate:
        type: string
      ID:
        type: integer
      RequestStatus:
        type: string
      RoadsResponse:
        $ref: '#/definitions/controllers.RoadsResponse'
    type: object
  controllers.RoadsResponse:
    properties:
      RoadImages:
        items:
          type: string
        type: array
      RoadNames:
        items:
          type: string
        type: array
    type: object
  controllers.response:
    properties:
      requestID:
        type: string
      roads:
        items:
          $ref: '#/definitions/ds.Road'
        type: array
    type: object
  ds.Road:
    properties:
      category:
        type: string
      endofsection:
        type: integer
      image:
        type: string
      length:
        type: integer
      name:
        type: string
      numberofstripes:
        type: string
      paidlength:
        type: integer
      price:
        type: integer
      roadid:
        type: integer
      speed:
        type: integer
      startofsection:
        type: integer
      statusroad:
        type: string
      trustmanagment:
        type: integer
    type: object
  ds.Travelrequest:
    properties:
      completiondate:
        type: string
      creationdate:
        type: string
      formationdate:
        type: string
      moderatorid:
        type: integer
      requeststatus:
        type: string
      travelrequestid:
        type: integer
      userid:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Road application
  title: Платные дороги
  version: "1.0"
paths:
...
  /roads:
    get:
      description: Get all roads
      operationId: get-roads
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.response'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Road'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Road'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Road'
      summary: Get Roads
      tags:
      - Roads
    post:
      consumes:
      - application/json
      description: create road
      operationId: create-road
      parameters:
      - description: road info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Road'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Road'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Road'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Road'
      security:
      - ApiKeyAuth: []
      summary: create road
      tags:
      - Roads
...
  /travelrequests:
    get:
      description: Get all travelrequests
      operationId: get-travelrequests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Travelrequest'
      security:
      - ApiKeyAuth: []
      summary: Get Requests
      tags:
      - TravelRequests
  /travelrequests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete travelrequest by ID
      operationId: delete-request-by-id
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Travelrequest'
      security:
      - ApiKeyAuth: []
      summary: Delete TravelRequest by ID
      tags:
      - TravelRequests
    get:
      consumes:
      - application/json
      description: Show roads by ID of request
      operationId: get-roads-by-id-of-request
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Response'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Road'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Road'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Road'
      security:
      - ApiKeyAuth: []
      summary: Get Roads by request ID
      tags:
      - TravelRequests
    put:
      consumes:
      - application/json
      description: Update travelrequest by ID
      operationId: update-travelrequest-by-id
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: request info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Travelrequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Travelrequest'
      security:
      - ApiKeyAuth: []
      summary: Update TravelRequest by ID
      tags:
      - TravelRequests
  /travelrequests/change-status-moderator/{id}:
    put:
      consumes:
      - application/json
      description: Update travelrequest by moderator
      operationId: update-travelrequest-status-by-moderator
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Travelrequest'
      security:
      - ApiKeyAuth: []
      summary: Update TravelRequest Status By Moderator
      tags:
      - TravelRequests
  /travelrequests/change-status-user/{id}:
    put:
      consumes:
      - application/json
      description: Update travelrequest status by user
      operationId: update-travelrequest-status-by-user
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Travelrequest'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Travelrequest'
      security:
      - ApiKeyAuth: []
      summary: Update TravelRequest Status By User
      tags:
      - TravelRequests
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"